docs: video.exists -> video.findById, 비디오 삭제 확인
exists는 true,false로 출력되기 때문에, true.owner 같은 요소는 없기에 undefiend가 나온다.
그래서 findById로 변경해서 객체를 찾도록 변경했다.

+++ src/controllers/videoController.js
@@ -50,7 +50,7 @@ export const postEdit = async (req, res) => {
       user: { _id },
     },
   } = req;
-  const video = await Video.exists({ _id: id });
+  const video = await Video.findById(id);
   if (!video) {
     return res
       .status(404)
@@ -103,7 +103,22 @@ export const postUpload = async (req, res) => {
 
 // * 비디오 삭제
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video) {
+    return res.status(404).render("404", {
+      pageTitle: "찾을 수 없는 영상입니다.",
+    });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
